data structure files added, not implemented yet. Fixing up Logger, needs to support wide strings for windows use. Working on JString.

+++ jackal/src/core/public/math/common.hpp
@@ -69,6 +69,6 @@ inline float Tanf(float value)
 
 inline float ToRadians(float degrees)
 {
-  return degrees * (J_PI / 180.0f);
+  return degrees * (static_cast<float>(J_PI) / 180.0f);
 }
 } // jkl
